pyproject.toml
[project]
name = "openbookqa-dspy"
version = "0.1.0"
description = "DSPy-based agent for OpenBookQA using OpenAI gpt-4o"
readme = "README.md"
requires-python = ">=3.13"
authors = [
  { name = "Your Name", email = "you@example.com" },
]
license = { text = "MIT" }
keywords = ["dspy", "openai", "huggingface", "openbookqa", "llm", "agents"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.13",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

# Runtime dependencies
dependencies = [
  "dspy-ai>=2.4.0",
  "openai>=1.43.0",
  "datasets>=2.20.0",
  "huggingface_hub>=0.24.0",
  "python-dotenv>=1.0.1",
  "pydantic>=2.7.0",
  "typer>=0.12.0",
]

[project.urls]
Homepage = "https://github.com/yourname/openbookqa-dspy"

[project.scripts]
openbookqa-dspy = "openbookqa_dspy.cli:app"

[build-system]
requires = ["hatchling>=1.22.0"]
build-backend = "hatchling.build"

[tool.uv]
# Dev dependencies are managed by uv
dev-dependencies = [
  "mypy>=1.10.0",
  "black>=24.8.0",
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "types-requests>=2.32.0.20240712",
]

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.mypy]
python_version = "3.13"
warn_unused_ignores = true
warn_redundant_casts = true
disallow_untyped_defs = true
disallow_any_generics = true
no_implicit_optional = true
strict_optional = true
check_untyped_defs = true
ignore_missing_imports = true
